requirements.txt
# MLOps Pipeline Dependencies

# Data & ML
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
pyarrow==14.0.1

# MLflow & Experiment Tracking
mlflow==2.9.2

# DVC - Data Version Control
dvc==3.30.1

# Feature Store
feast==0.34.1

# API & Web Service
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.3

# Monitoring
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0

# Airflow
apache-airflow==2.8.0

# Utilities
python-dotenv==1.0.0
joblib==1.3.2
pyyaml==6.0.1
requests==2.31.0

# Testing (optional)
pytest==7.4.3
httpx==0.26.0
